refactor(examples): 迁移 undo 至 plugins 目录
